requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
plotly>=5.18.0
wordcloud>=1.9.3
matplotlib>=3.8.0
